feat: deprecate <CardGrid /> for CSS (#30)
# Motivation

The grid might be another type of elements than a `div`. We notably need the grid to support `InfiniteScroll` too which happens to be a `ul`. For simplicity reason and to promote the usage of CSS rather than JavaScript properties, this PR deprecates `<CardGrid />` and introduces a new global style `.grid`

# Changes

Lib:

- remove `<CardGrid />`
- refactor style to global style `grid.scss`

Docs:

- dedicated grid docs