feat: theme, theme toggle and theme (#81)
# Motivation

Integrate the theme switcher, the theme toggle component and toggle component into the library.

# Note

The `theme.store` code moved is not the last version of NNS-dapp but previous one because we don't need the generic `writableStore` in the library. The store and related `theme.utils` were adapted to support the format of the property saved in `localStorage` - i.e. not a `string` but a `JSON.stringify(theme)` as currently on `main` branch of NNS-dapp.

# Changes

lib

- new `theme.store` and related utils (not exposed)
- component `<ThemeToggle />` and `<Toggle />` moved in the library

docs:

- related components pages added
- `Theming` doc updated
- add theme switcher to docs header with a popover