feat: new layout with a menu reveal and header positioning (#101)
# Motivation

Implement the new layout design that replaces the overlay menu with a reveal type menu and make the header sticky only on smaller devices in a backwards compatible way - i.e. implement the new layout in existing design.

In addition, drop javascript usage to make the menu sticky to a css only approach (🥳).