feat: backdrop over content if menu open on small devices (#113)
# Motivation

- improve user experience when the menu pushes the content on mobile devices
- improve user experience glitch when menu is open on small window, then user navigate and browser is resized and back button is still displayed

# Changes

- use the `<Backdrop />` component in the `<SplitPan />` but animate with CSS only and not with JavaScript / Svelte to have a smooth behavior without the need to observe window size with JS

# Edge case

If user as a small window, open the menu, make the window bigger, then navigate, then make the window smaller, the "back" button and menu open are going to be displayed. I think it's an acceptable edge case - i.e. a pure CSS solution is better than a JavaScript less smooth solution.