GIX-1864: Add property noPadding on Card (#293)
# Motivation

Fix the poor contrast on the token selector for ICP logo described here:
https://dfinity.atlassian.net/browse/GIX-1864

Solution: add a stronger outline to the logo when selected and on hover.

Yet, to be able to know when the card is hovered, we can't rely on the
hover of the content because the Card has some padding. Therefore, I'd
like to add a `noPadding` prop to the Card component. This way, the
child will manage the padding and can also manage the onhover of the
whole Card.

# Changes

* Add prop `noPadding` in `Card` component.

# Screenshots

![Screenshot 2023-09-12 at 09 30
51](https://github.com/dfinity/gix-components/assets/4550653/3415ac51-362f-4be1-94fe-71abe48426eb)