Chore: New colors for Tokens page (#319)
# Motivation

We are transitioning to a new set of colors defined in Figma. Not all
have been exposed yet. We do it gradually.

In this PR, I expose only the colors needed for the tokens page.

# Changes

* Add colors to light and dark themes.

# Screenshots

No new components, and the new colors are not yet used in any of the
current components.