fix(IDX-3073): allow Commit Formatting changes to trigger CI (#351)
# Motivation

IDX has been informed that when the CI job `Commit Formatting changes` is triggered, the PR is blocked by the CLA check workflow. This is because the github action `add-and-commit` intentionally does not trigger CI. See explanatation [here](https://github.com/EndBug/add-and-commit?tab=readme-ov-file#the-commit-from-the-action-is-not-triggering-ci):

```
GitHub sees that the push event has been triggered by a commit generated by CI, and doesn't run any further checks to avoid unintentional check loops.
```

However, because the `CLA Check` is required to run before the PR can be merged, the PR is blocked from merging. Previously this issue was resolved by creating an empty commit to re-trigger CI.

# Changes

This change adds a fine-grained Personal Access Token (PAT) to allow the `add-and-commit` action to commit to the PR and trigger CI. This is the solution described in their docs [here](https://github.com/EndBug/add-and-commit?tab=readme-ov-file#the-commit-from-the-action-is-not-triggering-ci).

I have solved this by extending the repo scope of `GIX_CREATE_PR_PAT` to include `gix-components` and granting the `sa-github-api` write access to the repository.

I have also updated the author to be github_actions instead of an empty author.

If you do not want these changes, feel free to close this PR, but you will need to resolve any CLA issues with the empty-commit solution described earlier.

# Screenshots

CI now runs on this commit.
<img width="949" alt="Screenshot 2024-01-11 at 13 19 48" src="https://github.com/dfinity/gix-components/assets/47304080/83ee9438-2f3f-4c28-9533-41629e9c3c5e">