feat: input type currency (#356)
# Motivation

I need more decimals in the input field of type icp and I don't want to ship a breaking change. 

So my first idea was to tweak the `inputType` set to `icp` but it leads to the issue with the scientific notation and I was concerned about making `icp` bind `value` of type `string` given that would break NNS dapp.

So ultimately I dediced to add a new `inputType` called `currency` which basically does the same thing as `icp` but, map the value to a `string`. That way we can avoid the issue of scientific notation. It's up to the consumer to use it correctly.

This PR also add a property to tweak the number of `decimals` accepted for input field type `currency`.

Both number of decimals and value parsed as number for `icp` remains that way the same.

# Changes

- add new type `currency` which basically does the same stuffs as `icp` but parse a `string`
- introduce an option `decimals` set to 8 per default which can be used with `currency`
- update few `value` prop in test reported as in error by webstorm (need string instead of number)
- update doc