build: fetch QRCode test video from dfinity download (#365)
# Motivation

Previous Google Drive hosting is for some reason broken and download required some logic anyway so we moved the file to dfinity download.

Also it looks like last PR was merged with failing test. Not sure if it was a glitch or recent changes in the pipeline would lead to this. We should keep an eye on it. As a results, this PR also update all required screenshots.

# Notes

Fee overlays background in light mode appears to have gone darker in light mode. I pinged David about it. Should not be a blocker for this PR which has for goal to fix the CI.