Feat: Add class utility to link to look like buttons (#362)
# Motivation

Add a utility class for the anchor tags to look like a primary button.

This is used in the Buy ICP flow of the NNS Dapp. It is also used with
an icon. Therefore, I also added the `with-icon` utility class to both
the link and button utility classes.

# Changes

* New class utility for anchor tag: `button` + `primary` + `full-width`
+ `with-icon`.
* Add `with-icon` to the button utility classes.
* LInk page docs.
* Add a card in Utilities linking to the link docs.
* Add e2e screenshot test for the links docs.
* Add docs about `with-icon` in the button docs.

# Screenshots

Check the screenshots in the PR.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>