Redo Tooltip logic (#368)
# Motivation

The current logic is a combination of patches and still the tooltip
isn't fully visible in several cases.
See for example "SpawnNeuronButton" in [this
document](https://docs.google.com/document/d/19nxAIC55W49gu-LDOLzJiLANs-vs8ObaudWl6aUy-AM/edit#heading=h.5yrpqsc8j2pm)
or in the nns-dapp itself.

Because I redid the logic, I'm hoping the individual arbitrary fixes are
no longer necessary. I went through all the tooltips in NNS dapp and
they all look good with this logic. But if we need to reintroduce any
logic we should do it with proper tests.

The new logic:
1. Tries to center the tooltip below the target.
2. If `top` is true, put the tooltip above the target instead of below
it.
3. If `center` is `true` skip steps 4 and 5.
4. If it causes the tooltip to overflow on the left, make the tooltip
touch the left edge of the container.
5. If it causes the tooltip to overflow on the right, make the tooltip
touch the left edge of the container.

I plan to also remove the `center` prop but that will be tested
separately in NNS dapp.

# Changes

1. Exported the component so it can be used from nns-dapp and other
clients.
2. Move the positioning logic to a separate util for easily testing.
3. Remove the `containerSelector` prop and instead discover the correct
container by finding the first ancestor with `overflow:hidden` or
`overflow-x:hidden`.
4. Don't assume a default scrollbar width if there is no scrollbar.
Because of the previous point we assume we actually found the scrollbar
if there is one.
5. Remove the bug to translate `0` instead of `-50%` when
`leftToMainCenter`.
6. Remove default off-center padding when not `leftToMainCenter`.
7. Add a checkbox on the showcase to toggle `overflow: hidden` to test
the container discovery.

# Tested

Unit tests added for the positioning logic.

Tested manually at
https://njmfy-reaaa-aaaaa-qabna-cai.dskloet-ingress.devenv.dfinity.network/

# Screenshots

<img width="413" alt="image"
src="https://github.com/dfinity/gix-components/assets/122978264/91bd772c-67fc-4db1-a7f7-a098bb9bd4dc">

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>