feat: improve mobile and desktop detection for QRCode reader mirroring (#378)
# Motivation

We received reports from users complaining about the QRCode reader being incorrectly mirrored on their devices. This issue was specifically noted on Samsung devices that support a stylus. The presence of a stylus leads the device to report to the user agent that it supports `any-pointer:fine`, i.e. it indicates support for a "mouse", which we previously used to detect desktop environments.

This PR revises our method for detecting mobile devices by utilizing the `NavigatorUAData` agent, which is available in Chrome. It also updates our approach for identifying desktop devices to correct the QRCode reader mirroring issue.

# Tradeoff

This implementation comes with a tradeoff. Since `NavigatorUAData` is not yet supported in Firefox, using this browser on a touchscreen device will result in it being detected as a mobile device.

This might be considered an edge case, given the current state of the APIs, this implementation represents our most accurate method for device detection available at this time.

# Resources

- https://caniuse.com/mdn-api_navigator_useragentdata
- https://developer.mozilla.org/en-US/docs/Web/API/NavigatorUAData
- https://stackoverflow.com/questions/11381673/detecting-a-mobile-browser

# Changes

- mirror qrcode reader if `isDesktop`
- review `isMobile`
- add TS support for experimental feature