NNS1-2966: Place tooltip directly in body (#406)
# Motivation

In NNS dapp we want to add a tooltip to every rounded number in the UI
to show the number in high precision.
Some of these numbers appear in the label of a `Checkbox` component,
which has `overflow:hidden`.
Currently this makes the tooltip completely invisible.
To avoid this, we have to place the tooltip element outside the
`overflow:hidden` element, for example in the document body.

# Changes

1. Place the tooltip element in the document body in `onMount`.
2. Remove the tooltip element in `onDestroy` to avoid getting more and
more elements in the body.
3. Use `onmouseenter` and `onmouseleave` instead of CSS to make the
tooltip appear, because the elements are no longer close enough in the
DOM to do this with CSS.
4. Increase the `z-index` of the tooltip to make sure it's not covered
by the backdrop when used inside a modal dialog.
5. Position the tooltip element when we show, because its position
relative to the body element might depend on the scroll position of
child containers. This also means that when you scroll while the tooltip
is shown, it may not track the tooltip target element. This should be OK
as the tooltip shouldn't be displayed for long.
6. Stop positioning the tooltip element on window resize because it is
now positioned when the tooltip is shown.
7. Remove `debounce` from tooltip positioning. It was there because it
could be triggered many times in a short time while resizing the window.
But we no longer trigger positioning on window resizing. Now that we
position the tooltip when it is shown, debouncing would cause a delay in
showing the window.
8. Remove the code to find the overflow container as we now always use
the body as the tooltip container.
9. Use `display:none` instead of `{#if}` to hide the tooltip when no
text is specified. Otherwise it may be moved back inside an
`overflow:hidden` container once text is defined again.

In the showcase:
1. Remove rerendering that was necessary to reposition the tooltip. Now
that the tooltip is positioned when shown, it's no longer necessary.

# Tested

1. Updated and added unit tests.
2. Updated e2e test to find the tooltip in the body.
3. Updated e2e test to hover a different button, demonstrating that the
tooltip is visible outside the `overflow:hidden` container.
4. Tested manually for every tooltip listed in [this
document](https://docs.google.com/document/d/19nxAIC55W49gu-LDOLzJiLANs-vs8ObaudWl6aUy-AM/edit#heading=h.2q2xotezucy0)
(which I wrote the previous time I updated the tooltip).

You can try it here:
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/

# Screenshots

See updated screenshot of e2e test in changed files.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>