Change tooltip colors to dropdown colors (#443)
# Motivation

In https://github.com/dfinity/gix-components/pull/393 I added colors
with the `--tooltip` prefix but these don't currently match the Tooltip
colors in Figma. They match the Dropdown colors in Figma, but they are
used for the `Popover` component. We also have Popup colors in Figma but
no Popover colors.
In the
[design](https://www.figma.com/design/x8dk5axQUf8Bi43uUsqCHc/NNS-Production-%26-Hand-over?node-id=1851-9872&t=YgFrTo6bb8sZXPJE-4)
for the `Hide zero balances` feature, the popup uses the Dropdown
colors. The implementation of the feature uses the `Popover` component.
I'm not sure all of this is correct but now at least our CSS matches
Figma regarding these colors.

I've also decided to just add all the Dropdown colors that are currently
in Figma rather than only the ones we use.

# Changes

1. Change names of CSS variables with `--tooltip` prefix to `--dropdown`
prefix.
2. Add additional `--dropdown` colors.
3. Update usage of these variables.

# Screenshots

unchanged

Tested manually in nns-dapp.