Add resetScrollPosition function (#444)
# Motivation

When switching between universes, the user would find themselves
somewhere in the middle of the list. This issue went unnoticed before
due to paginated loading, which resulted in the list being empty
immediately after switching. However, with actionables now being
preloaded, this has become a significant UX issue.

The proposals are rendered inside the `SplitContent` component. To reset
the scroll position without having to search through the DOM, we have
added a `resetScrollPosition` function.

# Changes

- new `resetScrollPosition` function.

# Screenshots

No visual changes. Tested in nns-dapp.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>