Make bottom logo element display:flex (#504)
# Motivation

Currently the logo gets an additional 3.781 pixels padding at the bottom
to make space for letter descenders.
We don't need or want this and making the element `display: flex` avoids
this.

With this change, the space below the bottom logo is exactly `24px`,
which matches the
[design](https://www.figma.com/design/x8dk5axQUf8Bi43uUsqCHc/NNS-Production-%26-Hand-over?node-id=4851-54807&t=HsQHPzgX8zjjd3Y0-4).

This also simplifies the planned change to remove the `MenuBackground`
component and make the full menu scrollable without hiding the bottom
logo. Because in NNS dapp, we currently have a fixed gap below the
GitHub button and theme switcher and we want to remove this and just let
the bottom logo element take up its own space. This is difficult to do
without affecting paddings if the element takes up a non-integer number
of pixels.

# Changes

Make the element that holds the bottom logo `display: flex`.

# Screenshots

See updated screenshots.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>