style: force content-box for Checkbox (#539)
# Motivation

It is useful to "force" the `Checkbox` component to have the CSS
property `box-sizing: content-box;`. That is because, if, for instance,
it is used in a context where the property is set as `border-box`, the
results are not so pretty:

![Screenshot 2024-12-05 at 19 38
11](https://github.com/user-attachments/assets/a8353936-d877-48f2-aa6e-79009370b9e1)


While, after this change, we have the intended results in all type of
contexts:

![Screenshot 2024-12-05 at 19 38
59](https://github.com/user-attachments/assets/14771630-0280-41db-85b7-ded84997480d)


# Side note

In OISY we have this exact example, but instead of changing it only for
OISY, I would imagine that this is the intention of the component, so I
changed it at the origin.

![Screenshot 2024-12-05 at 19 45
23](https://github.com/user-attachments/assets/846d3766-d8e4-45a4-a740-f50ba64b43d0)

should be

![Screenshot 2024-12-05 at 19 46
15](https://github.com/user-attachments/assets/dd5e8796-208b-4e46-9d34-7fef2ed5b0b1)

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>