feat: use an element target in infinite scroll (#569)
# Motivation

As I discovered in Juno (PR
[#1092](https://github.com/junobuild/juno/pull/1092)), where I have a
similar infinite scroll component to the one in Gix-cmp, the
`InfiniteScroll` component no longer works after the migration to Svelte
v5. The root cause of the issue is the following breaking change in
Svelte:

> beforeUpdate/afterUpdate no longer run when the component contains a
and its content is updated.


https://svelte.dev/docs/svelte/v5-migration-guide#Other-breaking-changes-beforeUpdate-afterUpdate-changes

Currently, we observe changes on the container to disconnect and
reconnect the intersection observer. However, since there won't be any
updates anymore, the observer will not reconnect, and the component will
not fetch any new data after the first scroll event.

This issue is confirmed by the E2E test I recently added, which is now
failing in the ongoing Svelte v5 PR (see this
[job](https://github.com/dfinity/gix-components/actions/runs/12984449192/job/36207364140?pr=548)
- `1) [Google Chrome] › e2e/infinite-scroll.e2e.ts:34:1 › should reach
end of scroll`).

To resolve and prevent this issue, we can change the approach by using a
target element to detect whether the element is intersecting.

# Changes

- Set an intersection observer on a targeted `<div>` to detect
intersection.
- Ensure the element is as hidden as possible. Using `content` or `none`
styles would prevent the observer from being triggered, as only visible
elements trigger the event.