feat: option to reset theme to system settings (#585)
# Motivation

In OISY, we would like to offer users the ability to select "System" in
addition to Dark or Light. Technically, this means reverting the theme
to its original value—removing it from local storage and deriving the
theme store value from the system settings.

# Changes

- Add a function to get the theme from the system settings
(`matchMedia`).
- Add a function to `applyTheme` but, delete the item in the
`localStorage`.
- Add function `resetToSystemSettings` to store.
- A bit unrelated (but not that much): we use the initial-value function
directly in the initialization of the store, instead of creating it
outside. I think it makes more sense since we do not want a static value
when creating this store, but we want to capture the current theme.

# Tests

Added a few tests for `themeStore` and `initThemeStore`.

---------

Co-authored-by: Antonio Ventilii <antonio.ventilii@dfinity.org>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Yusef Habib <yusef.fernandez@dfinity.org>