Hide top on scroll (#589)
# Motivation

To improve the mobile experience by maximizing screen space, this update
hides the header when the user scrolls down.

Initially, the approach was to hide the header on scroll down and show
it on scroll up. However, this caused glitches due to container resizing
because of SplitPage/padding: 0. Instead, a more reliable solution was
chosen: hiding the header when the scrollable content starts leaving the
visible viewport.

demo:
https://qsgjb-riaaa-aaaaa-aaaga-cai.mstr-ingress.devenv.dfinity.network/

# Changes

- Introduced a new ScrollSentinel component that observes when it leaves
the viewport and updates layoutContentTopHidden.
- The layoutContentTopHidden value is now used to toggle the header’s
visibility (to removed SplitPage top padding and the header when the
value is true)

- Tested on iOS and Android.

# Screenshots


https://github.com/user-attachments/assets/72df4b50-f0b6-4a80-862c-4b6f48fe95fd

### The glitch of the first approach


https://github.com/user-attachments/assets/e3b3c78e-7185-42a9-a92b-a4b40e127b35

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>