build: svelte v5 (#548)
# Motivation

We want to use the most recent tools, but we also want to avoid being
left behind, notably in terms of security. For example, SvelteKit is
going to patch a minor issue with cookies only in a future version.
That's why we are migrating to Svelte v5.

# Notes

This PR updates the dependencies and makes the required configuration
changes. No components are migrated since Svelte v5 is backward
compatible with how Svelte v4 is composed.

# Changes - App

- Bump Svelte v5 and related dependencies
- Set peer as Svelte v5

# Changes - Tests

- Bump Svelte Testing library
- Mock `ResizeObserver` and `element.animate` required by Svelte v5
- Update tests using `$on` and `$set` as documented
https://svelte.dev/docs/svelte/v5-migration-guide#Components-are-no-longer-classes
- Rename test files having to use `$state` for test purpose because
"rune is only available inside `.svelte` and `.svelte.js/ts` files"
- Remove `svelte/internal` config overwrite ("Your application, or one
of its dependencies, imported from 'svelte/internal', which was a
private module used by Svelte 4 components that no longer exists in
Svelte 5.")

# Changes - Screenshots

- The popover screenshots are modified because on `main` the code
snippet of the targeted page was actually not rendered correctly. A new
line was missing and the type of the property `export let button:
HTMLButtonElement | undefined` was not rendered. I did not managed to
fix that on main.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>