feat: Adds BottomSheet header footer and transition options (#628)
# Motivation

Adds the ability to have a `BottomSheet` transition in case we want to
slide it up as usually seen for bottom sheets.
Also adds a header and footer slot which can be used to place a header
or footer for the bottom sheet component.

There has also been a slight adjustment in max height as well as the
ability to make the content scrollable should it exceed the max height.

All defaults will behave the same as before to not affect existing
usages.

# Changes
- Added new slots and according styling
- Adjusted `max-height` and make it scrollable
- Added slide up transition and according prop to enable it
- Disabled border radius if not rendered as fix positioned card/sheet
(on desktop) since it cuts up text near the border
- Updated docs

# Screenshots

Default:

![image](https://github.com/user-attachments/assets/8294e0f0-b461-4460-9535-dbfdaf8eead8)

With header and footer:

![image](https://github.com/user-attachments/assets/7093cf5a-da5c-472e-bc69-90baa3ece623)

Only one of both:

![image](https://github.com/user-attachments/assets/4e4528af-7e4e-41c0-b0ab-f943304f64a2)

Desktop with slots:

![image](https://github.com/user-attachments/assets/9fc577a9-2ab6-4f10-81f2-ccbac2824af0)

Transition demo:
![Bildschirmaufnahme_20250502_154250-ezgif
com-video-to-gif-converter](https://github.com/user-attachments/assets/fce7d76f-8925-4f12-b86d-197ef46fdda8)

Scrolling working:

![image](https://github.com/user-attachments/assets/0834acaa-8572-481f-a36d-f7f6e49d6b9b)

Fixed slots stay fixed on scrolling:

![image](https://github.com/user-attachments/assets/e6549a88-c2c2-42a1-9497-8d2dcaf1da8c)

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>