feat: auto-generate docs for utilities (#642)
# Motivation

Today we added a utility in #640 that we are exposing because it can be
useful for consumer as well. Hence, I thought it would be nice to
automatically generate the documentation.

# Changes

-`tsdoc-markdown` **
  - Install dependency
- Create script to generate docs (target +page.md instead of a
README.md)
  - Add a script in package.json
- Add GitHub Actions to generate docs (scope it to only utils and page)
- Add new route '/utility-functions' and related menu entry

** Same setup as in IC-JS repo.

# Notes

The screenshots need to be regenerated because we are modifying the menu
with this PR.

# Screenshots

<img width="1536" alt="Capture d’écran 2025-05-17 à 13 47 48"
src="https://github.com/user-attachments/assets/aa336c60-935d-4103-bad3-2ba5ba59f084"
/>

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>