feat: Util for event modifier `preventDefault` (#664)
# Motivation

According to [Svelte 5
documentation](https://svelte.dev/docs/svelte/v5-migration-guide#Event-changes-Event-modifiers),
with the migration, the event modifiers are deprecated and should be
replaced by functions that have the same scope. For example, instead of:

```svelte
<button on:click|preventDefault={handler}>...</button>
```

One should do

```svelte
<script>
	function preventDefault(fn) {
		return function (event) {
			event.preventDefault();
			fn.call(this, event);
		};
	}
</script>

<button onclick={once(preventDefault(handler))}>...</button>
```

In this PR we create the `preventDefault` util, to help with the
migration (it will be used in other PRs).

# Changes

- Creating new util to prevent default actions, based on the Svelte
documentation.

# Screenshots

Created new tests, including a mock component.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>