feat: Migrate component `ScrollSentinel` to Svelte 5 (#663)
# Motivation

Migrating component `ScrollSentinel` to Svelte 5.

However, Svelte 5 is stricter on the `$state` types, meaning that
internal variable element could be `undefined` (not only `HTMLElement`).

To adapt to this change, we need to refactor a bit the `onMount` logic:
we should await for the DOM and the binding to finish through the Svelte
service `tick`.