fix: Avoid loss of precision in Input component (#678)
# Motivation

The Input component uses `Number()` when it is a currency type. However,
due to floating-point limitations of this method, the value that the
user inputs is different from the rendered ones sometimes, when the
decimals are too many (above 15-17), as explained in
[documentation](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number#number_encoding).
For example:

![Screenshot 2025-07-08 at 16 19
13](https://github.com/user-attachments/assets/90c612fa-0524-44e6-b44b-7c3b6ed041d3)

Unfortunately there is no real solution unless we create our own
`Number(value).toLocaleString()` method. Or if we use a common library
like
[decimals.js](https://www.npmjs.com/package/decimal.js?activeTab=readme).

We will opt for the second one, with a caveat: `Decimals` raises an
error for empty strings, or non-number strings, while `Number` would
just create the object `NaN`.

So, we make a safe-parsing of `Decimals`, and fallback to the "old"
method if it fails.

# Changes

- Install `decimals.js`.
- Change Input component to use Decimals but in a safe way (as explained
above).

# Screenshots

Added tests and tested manually:

![Screenshot 2025-07-08 at 16 24
00](https://github.com/user-attachments/assets/f60b7b94-3d4b-491a-afd0-917325ebb6c3)

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>