feat: Define CSS classes for levels of `Toast` component (#677)
# Motivation

Just an idea: it would be easier to customize the Toast style using CSS
styles, if we add the level among the classes.

For example, in OISY, if we want to customize the background of an error
toast we need to use the icon child:

```css
.toast:has(> .icon.error) {
	--overlay-background: var(--color-background-error-light);
	--overlay-background-contrast: var(--color-foreground-error-secondary);
}
```

While we could do:

```css
.toast {
	--overlay-background-error: var(--color-background-error-light);
	--overlay-background-contrast-error: var(--color-foreground-error-secondary);
}
```

So, we define custom CSS classes for each level of `Toast`: background
and foreground, inverted or not.

---------

Co-authored-by: David Dal Busco <david.dalbusco@dfinity.org>