chore: eslint v9 using lib eslint-config-oisy-wallet (#684)
# Motivation

Our all ecosystem is using eslint v9 except this project, therefore, we
also want to migrate it. To avoid maintenance overhead and to align code
patterns for the team, we want to use the
[eslint-config-oisy-wallet](https://github.com/dfinity/eslint-config-oisy-wallet)
rules.

# Notes

I did not realized there were so many changes but, it's mostly format.
It's a bit tricky to split that in many PRs, it would require a bit of
overhead of work so, if you are fine merging it as it, I would not be
against it but, I understand, it's meh to provide such a bulk PR even if
it's "just" eslint and format.

# Changes

- Install and setup lib
- Add two exceptions to the rules
- Update code according lint rules

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>