fix: Component `WizardModal` is back to work correctly (#695)
# Motivation

After migrating `WizardModal` to Svelte 5 in
https://github.com/dfinity/gix-components/pull/689 , it is not working
anymore using function `modal.next()`, `modal.back()` and `modal.set()`.


https://github.com/user-attachments/assets/ea0854ae-b0e5-486e-bcb3-d8605c1bcaf3

That is because the reactivity of `$effect` works different in the
following snippet:

```ts
// Svelte v4
$: ({ currentStep } = stepState);

// Svelte v5
$effect(() => {
  ({ currentStep } = stepState);
});
```
  
In Svelte 5, this reactivity is in fact triggered only when the entire
reference of `stepState` changes. However, if only its properties change
(like when we call functions `next`, `back` and `set`), then it does not
trigger a reaction, causing a stale component.

A possible solution is to directly trigger the update of `currentStep`
in each function.

We keep the `$effect` rune, just because it should be triggered on
assigning any value to `stepState` in any case.

# Changes

- Create small function to update the `currentStep` variable, taking
`stepState` as input.
- Use it both in the $effect rune and in all the exported functions:
`next`, `back`, `set`.

# Screenshots

I created new tests (with a test component). They work now, but it can
be easily verified that they were not working before this PR.

At the same here is a local manual test:



https://github.com/user-attachments/assets/072695c4-8219-44f9-99ce-886327592eb0